[Improvement][Seatunnel-web] Rename package com.whaleops.datasource to org.apache.seatunnel
